AxiosProvider useAxios refactoring
