+++ src/api/useGetUser.ts
@@ -1,6 +1,6 @@
 import { UseQueryResult, useQuery } from '@tanstack/react-query';
 
-import { useAxios } from 'providers/AxiosProvider';
+import { useAxios } from 'hooks/useAxios';
 import { useConfig } from 'providers/ConfigProvider';
 import { QueryKeys } from 'utils/constants';
